Perform int-like string key conversion when casting Hack arrays to PHP arrays
Summary:
Currently casting a dict or keyset to a PHP array does not perform
conversion of int-like string keys. This can result in PHP arrays that contain
both the integer and string version of a key, which is undesirable. So, change
this behavior and convert any int-like string keys to their int equivalents when
converting a dict or keyset to a PHP array. The downside of this is that such
conversions must now always require scanning the entire source dict or keyset.

Reviewed By: paulbiss

Differential Revision: D4500525

fbshipit-source-id: 884ed20bb82a9b331f790158cafc7122b68cfcb7