Refactor type_scan to produce addresses of pointers
Summary:
Getting locations of pointers is generally more useful:
- allows relocating unpinned objects
- HeapGraph could convert &ptr into an array index or property name

This diff adds a third pointer list in type_scan::Scanner
for addresses-of-pointers, and simply moves the dereferencing
from the generated scaner code to the client lambda of
Scanner::finish (a trivial change).

We still have the pointer list, because sometimes scanning will
produce pointer values that don't have an updatable address
(e.g. when the ptr is a key in a hashtable). Targets of such
pointers get pinned.

Also refactor the mark by generalizing CMark to Pin, and
update logging. We still track conservative-vs-exact scanning
(stats about the pointers we visit), and now log pinned_bytes
instead of ambig_bytes, to reflect how much of the heap was
pinned. The distinction will matter later when there are more
interesting reasons to pin objects besides conservative scanning.

Reviewed By: ricklavoie

Differential Revision: D4501334

fbshipit-source-id: 8de2d3455a6fab6b42085209348e484b473ca3c3