JSON RPC version of Find references response
Summary:
Currently, find references returns a list of (string, position).

The string tries to be some kind of an occurrence type description, for example, in:

  $group->fun();

the string might be "EntGroup::fun", or "EntGroupSubclass::fun" while the position points to "fun". As far as I know nobody cares about those strings (they were probably used for something in FBIDE).

Nuclide just takes the string out of first list element, ignoring the rest, and treats it as "referenced symbol name" (which is not really correct if the first result happened to be EntGroupSubclass while we were looking for references to EntGroup).

I'm changing the API to directly provide what Nuclide currently consumes - a single name, and a list of positions. The whole things is optional in case the thing under cursor is not a symbol.

Reviewed By: arxanas

Differential Revision: D4465076

fbshipit-source-id: 37d73c6e79e8be97732f1cb2e43efd5889f323be