Allow partial override for hhbbc bytecode hash/equals behavior
Summary:
Under certain circumstances, we may want to hash or compare bytecodes
with modified behavior for certain immediates.  For instance, if we
want to hash an entire CFG in a way that is invariant to block labeling
(but not to some canonical block ordering, e.g., RPO), we might want
to have all the usual hashing behavior, except overriding the typical
borrowed_ptr<php::Block> hash with a hash of the block's RPO ID.

This diff extends each bytecode op's hash() function to take an optional
hasher object, which can define an operator() on any subset of the
various bytecode immediate types.  It also adds an equals() function
with similar behavior for pairs of immediates.

Reviewed By: ricklavoie

Differential Revision: D4509208

fbshipit-source-id: cb5008baee2f1a9cdc9cf0d44044a56fd60e7080