Subtype logging
Summary:
Currently it's very hard to debug problems in the subtype and unification algorithms.
This diff introduces some logging, controlled by the insertion of calls to a new `hh_log_level` pseudo-function in Hack code.
Writing
```
hh_log_level(1)
```
enables logging of calls to a couple of subtyping functions whilst Hack is typing the code that follows, and
```
hh_log_level(0)
```
disables it again.

Reviewed By: jopamer

Differential Revision: D4522087

fbshipit-source-id: e1ac757b95b3024ddb339d615bc08044c274c43c