Forbid unset on non-end vec elements
Summary:
The current behavior of a vec possibly turning into a dict on unset is
confusing and unintuitive. Since nobody can really agree on what the proper
behavior is, just forbid it for now for non-end elements. Now performing an
unset on a non-end vec element will result in an exception being
thrown. Unsetting the last element in the vec is still allowed, as is "removing"
a non-present element (which is just a no-op).

Reviewed By: mxw

Differential Revision: D4514325

fbshipit-source-id: c91c36293b2cae1bc115e145df810554f0362eb6