Fix for assignment between newtype and lower-bounded generic parameter
Summary:
Suppose there is a definition
```
newtype N as C = ...
```
and in another file
```
function foo<T super N>(N $x): T {
  return $x;
}
```
Currently this is rejected, because the subtype checker eagerly chases the supertype of `N` as declared by `newtype`.
Instead, we need to try also the lower bound of `T`.

Reviewed By: dlreeves

Differential Revision: D4528541

fbshipit-source-id: 5f193ca308878e63b1378aa5690690cbd91a9274