Lex and parse "execution strings" in FF parser
Summary:
PHP supports strings of the form `command` where the command is executed. We do not support these in Hack, but to give better error messages and parse more files, we now parse them.

Not yet implemented: an error message if these are used in Hack. Also see notes in the lexer comments.

Reviewed By: eshrews

Differential Revision: D4516376

fbshipit-source-id: 2d8aa128176ee930d130356ec6f9e58b13ca9e6a