Add optional notices when binding a ref inside PHP arrays
Summary:
Add the ability to emit an optional notice when binding a ref inside a PHP
array. This is controlled by the new Eval.HackArrCompatNotices option. Future
notices of a similar manner will likewise be controlled by the same option.

Also fix a (probably theoretical) memory leak in APCLocalArray that can happen
if the operation on the escalated array throws.

Reviewed By: mxw

Differential Revision: D4520630

fbshipit-source-id: db914cb991bff762e2b2e90d1252503f401bab82