Make FF parser treat types as types in "new" expressions, and allow generics
Summary:
This diff fixes two problems with parsing "new".

First, we parsed the C in "new C()" as a simple name expression, when it is better characterized as a simple type.

If we wanted to search a parse tree for uses of the simple type "C", we'd certainly expect to find one here.

Second, we now allow "new C<int>()" as the HHVM parser allows. This will have to be added to the spec at a later date.

Reviewed By: andrewjkennedy

Differential Revision: D4516765

fbshipit-source-id: fb2b1b015e182f79e04a0a3d3af99549eee90b7a