Support explicit type arguments to new C(...)
Summary:
HHVM already has parser support for syntax such as `new C<int,string>(...)` and `foo<string,int>(...)`.

This change extends Hack with parser and type-checker support for the former: explicit type arguments to constructor invocations.
* If no explicit types are provided, then the behaviour is as today: Hack will generate fresh type variables for the arguments, which will unify appropriately depending on usage.
* If explicit types are provided between angle brackets, the number of type arguments must be correct (or else an arity error is produced).
* Any of the explicit types may be a special wildcard `_` type, for which Hack will generate a fresh type variable.
* This means that `new MyType<_>(...)` is equivalent to `new MyType(...)`.

The implementation is mostly straightforward. In the Named AST, the `CI` constructor in the `class_id` type now takes a list of type hints in addition to the static ID of the class. This will only appear (at present) as the first argument of the `New` constructor. Unfortunately, there is a lot of explicit pattern matching against `CI` in the codebase, and typically this can ignore the type hints.

Reviewed By: ericlippert

Differential Revision: D4507375

fbshipit-source-id: ea6c2b497624d046aefb59c0b39727381732c6f9