Move `Coverage_level.level` to `Ide_api_types.coverage_level`
Summary:
When adding the coverage levels request and response to `Ide_message`, there is a circular dependency. `Ide_message` depends on the `Coverage_level.level` type and calls a new function in `Coverage_level` to convert the coverage levels to an IDE response. But the IDE response type is defined in `Ide_message`.

The coverage level enum is just a type and doesn't depend on any of the functionality in either `Coverage_level` or `Ide_message`, so move it into a more basic module with a dependency on neither of these two modules.

This is used in a future commit.

Reviewed By: dabek

Differential Revision: D4495851

fbshipit-source-id: 838e49e726bff4d4f51850826c96e89615e1be8c