Add coverage levels request to JSON RPC
Summary:
The current coverage levels request is handled by calling `hh --color`. Currently, this returns its results in a weird JSON form: it provides a list of `text * coverage level` tuples. This is a bit weird and means that we have to transmit the entire file every time we request its coverage levels.

This diff moves this from the `hh` executable to the Nuclide RPC protocol, keeping the in the old form but deprecating it. It also adds the new form (a list of `range * coverage level` tuples`) to the JSON RPC protocol.

Reviewed By: dabek

Differential Revision: D4495868

fbshipit-source-id: 5037cedc2d6a6c4d636ad78c57ba199e328f7212