Better class constant propagation
Summary:
We were giving up on class constants which were defined in terms of
constants, or other class constants. Instead we can add dependencies
on the cinit function, and simulate it to evaluate constant values.

This wasn't a problem when hphpc was effectively doing this
optimization before we got to hhbbc, but with it turned off, it was an
8.5% regression in cpu time.

Reviewed By: swtaarrs

Differential Revision: D4513470

fbshipit-source-id: 0835051839d25138b3190bf3a7d9886a423460f0