Optimize unreachable code
Summary:
This breaks the Halt instruction into two new instructions: EndBlock
and Unreachable. EndBlock is treated exactly the way Halt was and is used for
tests and inline region selection, where we just need an artificial way to end
a block. Unreachable is used to indicate unreachable code in real IRUnits, and
we now optimize it aggressively. Any block that ends with an Unreachable should
itself be unreachable, so the rest of the block can be deleted.

We can also do things to effect-free jump instructions with at least one edge
going to an unreachable block, but those opportunities aren't yet showing up in
the code we generate. If and when that changes, it will be fairly easy to
strengthen our optimizations to match.

I also changed a few simplification functions to use Unreachable where
appropriate (it looks like they were written before the simplifier could handle
converting a non-terminal instruction into a terminal instruction).

Reviewed By: ricklavoie

Differential Revision: D4536797

fbshipit-source-id: 447263408b951a7f6d68d250805816dbe1cd0bd1