v0.4.4 - When expanding a state only bind rules on lines with overflow, or rules that those rules depend on
Summary:
If a line has no overflow, no need to break it up more, so find only rules that affect the first line with overflow and try binding those. In the case that a line must have overflow skip that line, and continue with the next.

This leads to massive performance improvements in files with large chunk groups as we are creating fewer states per each expansion. This has the twofold benefit of decreasing the number of states we need to sort, as well as reducing the branching factor of state expansion.

Differential Revision: D4485529

fbshipit-source-id: f478e0ec3c9e2e1857c76a89146709eb2f687d16