Add codegen option to full-fidelity-parse utility
Summary:
I have a full_fidelity_parse test utility for my own purposes, to which I have added the ability to do codegen.

This will be useful for my own testing purposes and was also a good way to learn what the interface to the codegen layer is.

Notice how I've organized the code to emphasize the steps that have to happen: parsing, naming, and codegen. I think this is pretty clear.

Reviewed By: hubyrod

Differential Revision: D4547554

fbshipit-source-id: 6aa69b0c6d6bb38c7810105fa90224ebdba56cd1