Support $ operator in FF parser
Summary:
PHP has a $ operator which can be used to dereference variables; eg:

    $y = 123;
    $x = "y";
    $w = "x";
    $a = $$x;  // 123
    $b = $$$w; // 123

We do not support this in Hack, but it would be nice if the FF parser could parse it so that we could have good error reporting, code fixes, etc.

Unfortunately the $$ feature makes it lexically ambiguous. Is $$$w to be lexed as $ $ $w, as it is in PHP, or $$ $w, as it is in Hack?

I've implemented a heuristic that is technically wrong in a few places, but we can revisit this decision at a later time and come up with a better heuristic.

Differential Revision: D4524281

fbshipit-source-id: 5b3f441aaa86b46c3d33cade5e9806d07a72c853