Enable hfsort based optimized translations
Summary:
Wires together parallel JITing, per-thread caches, and hf-sort. I found that since
hf-sort runs fairly quickly it does not need to be run in tandem with other phases
of retranslateAll as such we currently:

 - Construct the callgraph
 - JIT all optimized translations into cache space set aside
 - Sort the translations
 - Relocate the translations

This will allow us to eventually modify the JIT to update the callgraph as it
generates translations, dropping inlined calls and adding calls detected at
JIT time.

Reviewed By: swtaarrs

Differential Revision: D4246075

fbshipit-source-id: 7ebdf87d2ff7463c4ee9aa0b622b03fbea97ef7f