Speed up lookup_param_prep
Summary:
I noticed that during the analyze steps, hhbbc was spending > 20% of
the time in lookup_param_prep. The reason is that there are often many
methods with the same name, and in that case we end up iterating all
the methods for every parameter, and then, in almost all cases, we
return PrepKind::Val.

This precomputes a map from method names to a bitmap indicating which
parameters might be by ref, and short circuits when that tells us that
it can't by by ref.

Reviewed By: alexeyt

Differential Revision: D4552100

fbshipit-source-id: 80c1aada895fe407d2a2f3495fce50cb1406b30e