Fix lexing of escaped braces in double quoted strings
Summary:
The rules for escaped braces in strings is even more bizarre than I thought.

I thought that the rule was that "\{$x...." simply meant that this was not an escape sequence, the backslash was part of the string value.

What it actually means is, yes, the { is escaped, *AND* the \ is part of the string value.

So "\{$x,$y\}" is \{123,456\}, and not, as you might think, {123,456}.

I've fixed the lexer so that the \{ causes it to not consider the { as starting an embedded expression.

Reviewed By: dlreeves

Differential Revision: D4533939

fbshipit-source-id: bc490bdfd5c429a416694681dce60ff88436f841