Speeding up the hphpd_auth_token() script
Summary:
This modifies the script which powers the output of the `hphpd_auth_token()` extension. Previously, we relied on retrieving the output of `php scripts/login/ldap_from_devserver.php`. Now, we are going to call the script located at `/usr/local/bin/get_session_token_from_devserver`, which was written by azich.

This script is not a php script and hence can be directly executed, as opposed to running `exec("php", <path>)`, which is the main change here.

The chef diff which put the new script on all devservers: D4553298

Reviewed By: paulbiss

Differential Revision: D4576908

fbshipit-source-id: 4e4e65752c32640f7e765eddabd1ccde12652240