Allow right braces inside XHP bodies in FF parser
Summary:
The FF parser was choking on <x>}</x> because it was assuming that the right brace was the end of a braced expression, which is unexpected at this point.

I've modified the parser so that it treats this as an embedded right brace token, rather than as text. This is a likely mistake, and we will probably wish to create a warning about it at some point.

Reviewed By: andrewjkennedy

Differential Revision: D4538313

fbshipit-source-id: debaaea9cc141adc9265025750874d73a0b717d6