Fix a bug handling VarEnvs
Summary:
When assert was called with an ActRec (ie not via FCallBuiltin) it
would copy the VarEnv from the callee to assert's own frame, because
invokeFunc assumed that the VarEnv it was passed was attached to the
current ActRec.

The trouble with that is that assert's own locals (its parameters)
were injected into the environment (which is observable in itself);
but worse, if the calling function called assert more than once the
incoming VarEnv would clobber the params with their values from the
previous call.

This rewrites the VarEnv code to just assume that its inherited from
an outer scope, and that none of the intervening frames have VarEnvs.

Reviewed By: ricklavoie

Differential Revision: D4577572

fbshipit-source-id: 8b717094e8498efafdc905e9dec631b6a550cde6