Fix a bug in createBacktrace
Summary:
If the top frame was a builtin, we didn't report file/line number
information, so whether we got file/line number information depended
on whether we used FCallBuiltin.

Fix it to look through to the first user file and report that - which
appears to match zend's behavior (I tried array_map -> array_map ->
sort).

Reviewed By: paulbiss

Differential Revision: D4579292

fbshipit-source-id: a6ca301f33e905abcffc86f1b546e0d32217d31a