Fix an issue with builtin default arguments
Summary:
Various builtins use KindOfUninit as a flag to indicate the parameter
was not passed, and can have significantly different behavior when
called this way vs with null.

When emitting FCallBuiltin, the emitter did the right thing, but for
NativeImpl it did not.

The new unit test failed because it went through the native stub, and
saw null for the unpassed argument, rather than uninit.

This showed up when turning off FCallBuiltin in the emitter.

Reviewed By: ricklavoie

Differential Revision: D4574999

fbshipit-source-id: 21946fdca98c00da50e66be5aea98f876822d813