Fix order of destruction for FCallBuiltin args
Summary:
When a parameter failed to coerce, we destroyed the arguments in the
reverse order compared with a normal call - this meant that the order
of destructors was different depending on whether FCallBuiltin was
enabled.

Make everything match the case where FCallBuiltin is disabled (which
also matches PHP5 and PHP7).

There were two almost identical tests, which from their names were
supposed to test the two cases where we side exit while checking
params, and throw while checking params. But both of them tested the
side-exit path; so I fixed the one that was obviously supposed to
throw to actually throw.

Reviewed By: ricklavoie

Differential Revision: D4581541

fbshipit-source-id: a30100f8ec197a68b58432b536237cb1e703d32d