Make Debugger Server bind to localhost by default, and allow it to be configurable
Summary: In D2220234 we added options so that the admin server and satellite servers could be bound to specific addresses. I'm adding an option to support the same functionality for the debugger server, which today binds to all interfaces with no option for configuration. I'm also changing the default behavior so that it will bind to localhost rather than all interfaces when enabled.

Reviewed By: binliu19

Differential Revision: D4570745

fbshipit-source-id: 68cf85af4131967f49d16d1c552bdfa5a380eef1