v0.4.7 - Modify state expansion to prevent making a duplicate state
Summary:
Imagine each state is represented by a bit vector of rule values (N means unbound):

Previously state expansion (for a solution with two independent rules) went like this:

```
N N -> 1 N , N 1
1 N -> 1 1
N 1 -> 1 1
```

Now it works like this:

```
N N -> 1 N, 0 1
1 N -> 1 1
0 1 -> {}
```

removing the generation of duplicate states. There is also some logic to not bind the rule to 0 for future states if we were unsuccessful in creating a rule-value-map that bound the rule to 1.

Reviewed By: jamesjwu

Differential Revision: D4486116

fbshipit-source-id: ba4e696e51d071d0574acffa853c3a5f83d19d51