v0.4.8a - Refactor adding lines field to solve state type
Summary:
This diff adds a pre computation of a "lines" structure that is useful when calculating other properties of the solve_state.

"Lines" is a list where each element is a tuple (int, ISet.t) that represents the calculated overflow and set of rules that apply to a line of the outputted state.

This intermediate state can be used to calculate total overflow of the state, as well as candidate rules when expanding the state, without having to iterate over all of the chunks multiple times.

Differential Revision: D4529156

fbshipit-source-id: e83f6c7203611408016db95a2c4fc238c0866992