array_column allows a null $column_key
According to php docs: http://php.net/manual/en/function.array-column.php

column_key
The column of values to return. This value may be an integer key of the column you wish to retrieve, or it may be a string key name for an associative array or property name. It may also be NULL to return complete arrays or objects (this is useful together with index_key to reindex the array).