connectAndQuery API
Summary: clients sometimes create connections and then don't use them for a while which results in the connections timing out. In simple cases where the client wants to execute a single set of queries on a connection and don't need a persistent connection we can give them something that will create the connection and execute the query immediately after.

Reviewed By: anca-agape

Differential Revision: D4544949

fbshipit-source-id: 5b46823106a566191a957a813a03e3f2e2e38e49