Handle TInitGen properly in HHBBC's Hack array type functions
Summary:
In a few situations in HHBBC, we may attempt to "set" a TInitGen inside a Hack
array. One example is if we're analyzing a FPassM hhbc. We don't know if the
function we're calling takes the param by ref or not, so we analyze it both as a
CGetM and a VGetM. The VGetM analysis will attempt to simulate the type effects
of putting a TInitGen inside the Hack array.

Unfortunately the Hack array type functions didn't handle TInitGen properly, as
it would simply union it in with the existing types, causing invariants to be
broken. Fix this by first checking if the value type could be a TInitCell. If
not, then we have TBottom because this operation will always throw. Otherwise,
we can proceed as if the type is a TInitCell because we cannot statically
determine whether the operation will throw or not, but if it does not, its a
TInitCell. We can only safely assert the operation will not throw if we
definitely know the type is a TInitCell.

Reviewed By: markw65

Differential Revision: D4604448

fbshipit-source-id: f8039cda6aaade6c491288b3f2d7dd367b4300b9