Fix a TERRIBLE performance bug in incremental mode
Summary:
Incremental mode in a nutshell: calculate new definitions, compare them with old, infer dependencies that need rechecking based on this comparison. There is a bunch of logic that goes into comparing definitions in a "smart" way. One of them is Decl_pos_utils.NormalizeSig, which replaces all of the positions by "Pos.none", so that definitions that only differ by position are considered the same (for some applications).

Here is the fun part: we added more positions to some of the types, without updating this function, so incremental mode often ends up (incorrectly) telling you that files need to be rechecked when they don't.

I'm fixing one case that has clear repro here, but there is much more suspicious stuff here - will file a task to come back to it later.

Thanks hannesr for providing a good repro case!

Reviewed By: oulgen

Differential Revision: D4603999

fbshipit-source-id: 2a7022a961e8e6bfc18dd6f76d149afdff48296a