Implement && and ||, helper methods for instructions
Summary:
I've implemented && and ||.

While doing this I noticed that there was an enormous amount of "ceremony" in every use of the "instr" method.  Lots of module names and parentheses that were obscuring meaning. I've started adding helper methods to make our code gen much easier to understand; it now closely resembles the actual codegen.

Let me know if you like this, and consider using it.

Reviewed By: oulgen

Differential Revision: D4609101

fbshipit-source-id: 0f94ac88725dbe822c75f666124b510fcd149224