First cut at implementing pipe operator
Summary:
An initial implementation of the pipe operator. The $$ is not yet lowered to its form as a local, but everything else is in place.

I'd rather not have to thread state through the expression generator remembering that $$ means a particular local in a particular context. It will be easier to emit a dummy instruction for uses of $$ and then rewrite them post-hoc. I'll do that in the next diff.

Reviewed By: oulgen

Differential Revision: D4609650

fbshipit-source-id: 9b0da5f7d3242e5c9c9ac25de87785c1381519ef