Codegen: post-codegen relabelling
Summary:
We introduce a post-codegen remapping of labels in order to more closely match the behaviour of HHVM. The emitter is now free to generate fresh labels in any order it pleases, and to define multiple labels at the same instruction, and to define labels that aren't referenced. Furthermore, even if the label supply is global, the final labelling will be
unique to each method body.

In essence, we relabel instructions such that
1. No instruction is preceded by more than one label;
2. No label is unreferenced;
3. References to labels occur in strict label number order, starting at 0.

Exceptions and fault blocks are not handled; these will be dealt with in a future diff.

Reviewed By: ericlippert

Differential Revision: D4611769

fbshipit-source-id: 920de25f90fe71c386c06bf7ada1ddd3e1c38920