foreach loop codegen
Summary:
Basic foreach loop codegen is now working, but the key and value variables must be local variables. We support exotic loops like "I've got an associative collection of pairs that I wish to destructure":

    foreach($a as $key => list($x, $y))

We don't do codegen for that yet.

Also, I don't yet know under what circumstances the W and M forms of the instructions are generated.

Reviewed By: oulgen

Differential Revision: D4613573

fbshipit-source-id: 687b0455f69588fd1572cfb3db2f00ffbbe9dda7