Allow fpi regions to end on a terminal
Summary:
hhbbc tried to avoid causing issues with fpi regions by not stripping
dead code inside them. But that only worked as long as the rest of the
fpi region was contained in the same block as the first unreachable
instruction. If it wasn't, the unreachable blocks were stripped
anyway.

The new testcase produces code that doesn't pass the verifier, and
it throws while there's a pre-live actrec on the stack, but no fpi
region to protect it.

This fixes the hhbbc emitter to properly close fpi regions in the
presence of unreachable code, and fixes the optimizer to strip
unreachable code regardless.

Reviewed By: edwinsmith

Differential Revision: D4617939

fbshipit-source-id: 0bd132d5adc5c2fd496fce4ffb26976cf2c76b77