Tools for generating visitors for AST(X) data structures
Summary:
After offline discussion of the RFC diff D4501400, we decided that it would be useful to check in the scripts used to generate visitors for the AST data structure.

These are the tools //other than// the [[ https://gitlab.inria.fr/fpottier/visitors | visitors package on opam. ]]

The (revised) RFC diff will be stacked on this one and the diff eliminating the old visitor will be stacked on that. These diffs will land as a stack.

Reviewed By: andrewjkennedy

Differential Revision: D4516344

fbshipit-source-id: 9e46857dac8ccbba0c22b15eb49ecb3132cf0e22