Put instanceof on classname<T> under experimental instanceof and improve error message
Summary:
The new safe `instanceof` feature on generic classes is currently enabled only if `instanceof` appears in the `enable_experimental_tc_features` in `.hhconfig`, or under `hh_single_type_check`.
But `instanceof` on `classname<T>` is rejected even if this experimental feature is not enabled. This has caused build breaks in open-source projects. So for now let's put it under the
experimental feature guard too.

Also, the error message is confusing, so let's take the opportunity to improve it here.

Reviewed By: dabek

Differential Revision: D4617688

fbshipit-source-id: d1a29132423ced9beb3fe1a6d1476d378958f464