Codegen for try-finally with continue and break, part seven
Summary:
The plumbed-through continue-and-break-label state is now removed. Continue and break are realized as dummy instructions, which are then immediately rewritten by a transformation pass.

Gear is in place for handling multi-level break and continue, which PHP supports but Hack does not.

I am now in a position where I can begin to refactor try-catch-finally to support codegen for finally-blocked continue and break.

Reviewed By: oulgen

Differential Revision: D4623920

fbshipit-source-id: b9901ef958efa8e0d17422c0169dd0bd9f889cd2