Codegen for try-finally with continue and break, part eight
Summary:
Beginning to simplify the try-catch-finally code.  We begin with the observation that

    try { } catch { } finally { }

and

    try { try { } catch { } } finally { }

should have the same code generated. Therefore we never need to write a try code generator that has to deal with *both* catch and finally in the same try.

I have changed the call site so that it never passes both catch and finally.  This unfortunately has a side effect in the label generator which is not canonicalized by Andrew's label canonicalizer.  We'll fix the label canonicalizer in a later diff.

In the next diffs I'll tease the existing code which generates try-catch-finally apart into methods that deal just with try-finally and just try-catch.

Reviewed By: oulgen

Differential Revision: D4624302

fbshipit-source-id: 658eb328d7b2c9fdf3cba8d37a71abac66d6f197