Add CLI server to hhvm for scripts
Summary:
Adds a unix socket service for executing PHP scripts. The main motivation here
is to allow scripts to take advantage of the TC but still execute in a context
where they can read/write/exec as the user that invoked them. By default the
server is enabled but the client needs to opt in to using it. As there are
some inconsistencies with this execution mode I didn't want to default to it.

See cli-server.h for more details.

Reviewed By: swtaarrs

Differential Revision: D4483942

fbshipit-source-id: 5f0879dc7dbbbc7bf5e4a7dc6b8738da2fe97115