Cleaning up the lowerer for maintainability (and perf)
Summary:
This cleanup was long in the works and this yack is now properly bold.

The `env` has been dressed down to the "bare necessity". This could be completely eliminated, but we're well past the point of yack shaving.

The one remaining thing to do is to fix string infidelity (spaces being handled incorrectly in strings inside XHP).

Reviewed By: ericlippert

Differential Revision: D4605873

fbshipit-source-id: 233ed9079c5a195bb67f2bebae4d81041b2fbd04