Expanded visitors for AST
Summary:
Tagging especially hubyrod and andrewjkennedy for the codegen part. Anyone else should feel free to jump in.

I've used [[ https://gitlab.inria.fr/fpottier/visitors | François Pottier's new OCaml syntax extension ]] to generate more visitors for the AST type. This diff is stacked on top of the diff of the scripts used to generate this. The shame should be done for the newly functorised N/TAST.

This code is (mostly; `ast_visitors_ancestors.ml` excluded) generated by Pottier's extension. Considering the OSS dependencies and the issues checking things into the repo, I'm proposing checking in this generated code, //without// the generator as a ThirdParty dependency (instead, for reference, it goes in `hphp/hack/tools/visitors`).

Because I wanted one visitor for the entire AST, I changed `ast.ml` by replacing all-but-the-first occurrence of `type` with `and`. This way, all types in the AST are one group of mutually recursive (as far as the syntax extension is concerned) types. Finally, I added
```
[@deriving visitors { variety = "VISITORFLAVOUR" ; ancestors = ["VISITORFLAVOUR_base"]}]
```
to the end of the last type's definition.

Reviewed By: andrewjkennedy

Differential Revision: D4501400

fbshipit-source-id: 9aa17d0c2710a9111d802cdf2fba49304cfc871f