Clean up extract-decl-vars, add unique-list ADT.
Summary:
Killing two birds with one stone here. First, I noticed that the interface to extract_decl_vars was very baroque. It should do one job: take a sequence of instructions, return a list of unique local variables. But to do that it required a set and a list, it took a list of instructions instead of an instrseq, and its implementation was a re-invention of fold_left.

Second, I'm going to need the "list of unique items sorted in stable, predictable order" feature for a list of jumps that are finally-blocked, so I've generalized it to its own functorized module.

Reviewed By: oulgen

Differential Revision: D4632290

fbshipit-source-id: a898588e851263a5c343dd4150c3a95980e8bbe6