Entry-point for the FFP+Lowerer, in prep for the cut-over
Summary:
This diff prepares the Full Fidelity Parser and the lowerer for the cut-over in our binaries. Mostly this is a rename and a polish of the entry-point.

(Having done an `arc pull`, I fear I have conflated a sync-fix for `DictionaryTypeSpecifier` with this diff also, but it's minor. If need be, I'll split it out.)

Reviewed By: hubyrod

Differential Revision: D4620084

fbshipit-source-id: 74db360ccafbb48edb52563c4b13fc8999f29a3f