Add original level tracking to synthetic continue/break instrs
Summary:
The "continue" and "break" instructions are rewritten gradually into jumps.  It turns out that for error reporting purposes we will need the "original" level of the break/continue statement, so we preserve that fact in the pseudo-instruction.

This information will be consumed in an upcoming diff.

I've also moved these to their own instruction sub-type, since (1) they are quite different, not being "real" instructions, and (2) in a later diff I'm going to need to do a match on just these two possibilities.

Reviewed By: hubyrod

Differential Revision: D4633913

fbshipit-source-id: eccce2c53efdc463f3c780503708486a56bb58f7