Remove incorrect rewrite of continue/break in try-finally
Summary:
The correct rewriting of a continue / break in a try-finally is to (1) set a local indicating that there needs to be a branch after the finally, and (2) branch to the finally, and (3) take that branch at the end of the finally if necessary.

The existing codegen rewrites continue / break in a try-finally does only (2), and it's going to be confusing to have that rewrite there while I'm working on putting together the correct codegen algorithm.

I've therefore taken it out.  Finally-blocked continue / break is now generated as though it was not finally-blocked. This is wrong, but the code was already wrong, and it will be easier making it right starting from a blank slate.

Reviewed By: hubyrod

Differential Revision: D4633955

fbshipit-source-id: 83a784dee31ded4b12d880e5730e85d4b33e135e