Make try-catch a normal instruction by splitting it up
Summary:
We have the oddity that a try-catch instruction is a label plus a list of instructions that go in the try block, but this is unnecessary. All we need to know to generate correct code is where the try block begins and ends, and we can do that with two instructions.

We won't be able to get rid of this oddity for fault blocks entirely, but I'll simplify them a bit in the next diff.

Reviewed By: oulgen

Differential Revision: D4634936

fbshipit-source-id: 91ebda6f1525dc3292988a11c3b2259de725277b