Remove unnecessary branch from try-finally
Summary:
A try-catch must have a jump at the end of the try block that skips the catches. But we were also generating a jump at the end of a try-finally. Since the finally immediately follows the try body, this is a jump to the next instruction and can be eliminated.

Note that we will eventually need a label to jump to the finally for finally-blocked continue and break, which, again, is the whole point of this now 21-deep stack. Soon!

I've also eliminated the weird unsets in fault blocks until we understand what they are for.  I suspect they are bugs in HHVM.

Reviewed By: oulgen

Differential Revision: D4635227

fbshipit-source-id: 5c5fe8ea00df78c73b2a91a9ee1f5bc400870bcb